backend/services/friends-chat-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
redis==5.0.1
python-multipart==0.0.6
pytest==7.4.3

